member approval

+++ routes/member.py
@@ -13,7 +13,7 @@ from models.shares import MemberShares as MemberSharesModel
 class Member(Resource):
 
     @token_required
-    @user_check(user_type=['teller'])
+    @user_check(user_type=['teller', 'approver'])
     def get(self, uuid):
         member = MemberModel.query.filter_by(uuid=uuid).first()
         result = member_schema().dump(member)
@@ -93,35 +93,38 @@ class Member(Resource):
             print(e)
             return make_response({'error': 'Something went wrong'}, 500)
 
+    @token_required
+    @user_check(user_type=['teller', 'approver'])
     def put(self, uuid):
         params = request.get_json(force=True)
         userInfo = decode_token()
+        userType = userInfo.get('user_type')
 
         memberData = params.get('memberForm')
-        del memberData['shares']
+        if userType == 'teller':
+            del memberData['shares']
+            del memberData['loans']
 
-        forDeletion = params.get('forDeletion')
-        beneficiaries = None
-        share = None
+            forDeletion = params.get('forDeletion')
+            beneficiaries = None
+            share = None
 
-        if memberData.get('share'):
-            share = memberData.get('share')
-            del share['total_share_amount']
-            del memberData['share']
+            if memberData.get('share'):
+                share = memberData.get('share')
+                del memberData['share']
 
-        if len(memberData.get('beneficiaries')) > 0:
-            beneficiaries = memberData.get('beneficiaries')
-            del memberData['beneficiaries']
-        else:
-            del memberData['beneficiaries']
+            if len(memberData.get('beneficiaries')) > 0:
+                beneficiaries = memberData.get('beneficiaries')
+                del memberData['beneficiaries']
+            else:
+                del memberData['beneficiaries']
 
+        print(memberData)
         updateMemberData = {
             **memberData,
             'updated_at': datetime.now(),
             'updated_by_id': userInfo['id']
         }
-
-        print(updateMemberData)
 
         try:
             # Check if there are other member with the same set of details
@@ -141,45 +144,47 @@ class Member(Resource):
                 MemberModel.query.filter_by(
                     uuid=uuid).update(updateMemberData)
                 db.session.commit()
-                if beneficiaries:
-                    memberInfo = MemberModel.query.filter_by(uuid=uuid).first()
-                    dbs = db.session
-                    forCreate = []
-                    for beneficiary in beneficiaries:
-                        if beneficiary.get('id'):
-                            forUpdate = {
-                                **beneficiary,
-                                'updated_at': datetime.now(),
-                                'updated_by_id': userInfo['id']
-                            }
-                            del forUpdate['id']
-                            BeneficiariesModel.query.filter_by(id=beneficiary.get('id')).update(forUpdate)
-                        else:
-                            benefitParams = {
-                                **beneficiary,
-                                'member_id': memberInfo.id,
-                                'created_at': datetime.now(),
-                                'created_by_id': userInfo['id']
-                            }
-                            forCreate.append(
-                                BeneficiariesModel(**benefitParams)
-                            )
-                    dbs.bulk_save_objects(forCreate)
-                    dbs.commit()
 
-                if forDeletion:
-                    dbs = db.session
-                    for beneficiary in forDeletion:
-                        BeneficiariesModel.query.filter_by(
-                            id=beneficiary.get('id')
-                        ).delete()
-                    dbs.commit()
+                if userType == 'teller':
+                    if beneficiaries:
+                        memberInfo = MemberModel.query.filter_by(uuid=uuid).first()
+                        dbs = db.session
+                        forCreate = []
+                        for beneficiary in beneficiaries:
+                            if beneficiary.get('id'):
+                                forUpdate = {
+                                    **beneficiary,
+                                    'updated_at': datetime.now(),
+                                    'updated_by_id': userInfo['id']
+                                }
+                                del forUpdate['id']
+                                BeneficiariesModel.query.filter_by(id=beneficiary.get('id')).update(forUpdate)
+                            else:
+                                benefitParams = {
+                                    **beneficiary,
+                                    'member_id': memberInfo.id,
+                                    'created_at': datetime.now(),
+                                    'created_by_id': userInfo['id']
+                                }
+                                forCreate.append(
+                                    BeneficiariesModel(**benefitParams)
+                                )
+                        dbs.bulk_save_objects(forCreate)
+                        dbs.commit()
 
-                if share:
-                    shareId = share.get('id')
-                    del share['id']
-                    MemberSharesModel.query.filter_by(id=shareId).update(share)
-                    db.session.commit()
+                    if forDeletion:
+                        dbs = db.session
+                        for beneficiary in forDeletion:
+                            BeneficiariesModel.query.filter_by(
+                                id=beneficiary.get('id')
+                            ).delete()
+                        dbs.commit()
+
+                    if share:
+                        shareId = share.get('id')
+                        del share['id']
+                        MemberSharesModel.query.filter_by(id=shareId).update(share)
+                        db.session.commit()
 
                 return {'response': True}
             return make_response({'error': 'User is already made'}, 500)
@@ -191,7 +196,7 @@ class Member(Resource):
 class MemberList(Resource):
 
     @token_required
-    @user_check(user_type=['teller'])
+    @user_check(user_type=['teller', 'approver'])
     def get(self):
         try:
             members = MemberModel.query.order_by(MemberModel.status.asc()).all()
